Call to get sales center is passing wrong value for the community
